- move actionTypes into constants directory
